cfastbot: make cfastbot clearer and more robust

+++ Cfastbot/run_cfastbot.sh
@@ -1,5 +1,109 @@
 #!/bin/bash
 
+# CFASTbot
+# This script runs the CFAST verification/validation suite 
+# on the latest revision of the repository.
+
+#---------------------------------------------
+#                   CHK_REPO
+#---------------------------------------------
+
+CHK_REPO ()
+{
+  local repodir=$1
+  
+  if [ ! -e $repodir ]; then
+     echo "***error: the repo directory $repodir does not exist."
+     echo "          Aborting smokebot."
+     return 1
+  fi
+  return 0
+}
+
+#---------------------------------------------
+#                   CD_REPO
+#---------------------------------------------
+
+CD_REPO ()
+{
+  local repodir=$1
+  local branch=$2
+  
+  CHK_REPO $repodir || return 1
+
+  cd $repodir
+  if [ "$branch" != "" ]; then
+     CURRENT_BRANCH=`git rev-parse --abbrev-ref HEAD`
+     if [ "$CURRENT_BRANCH" != "$branch" ]; then
+       echo "***error: was expecting branch $branch in repo $repodir."
+       echo "Found branch $CURRENT_BRANCH. Aborting smokebot."
+       return 1
+     fi
+  fi
+  return 0
+}
+
+#---------------------------------------------
+#                   usage
+#---------------------------------------------
+
+function usage {
+echo "Verification and validation testing script for cfast"
+echo ""
+echo "Options:"
+echo "-a - run automatically if cfast repo has changed"
+echo "-c - clean cfast and FDS-SMV repos"
+echo "-f - force cfastbot run"
+echo "-h - display this message"
+echo "-i - use installed smokeview and background (if using the 'none' queue)"
+echo "-I - compiler [ default: $compiler]"
+echo "-k - kill cfastbot"
+echo "-m email -  email_address "
+echo "-q queue_name - run cases using the queue queue_name"
+echo "     default: $QUEUE"
+echo "-s - skip matlab and guide generating stages"
+echo "-u - update cfast and FDS-SMV repos"
+echo "-U - upload guide (only by user: cfastbot)"
+echo "-v - show options used to run cfastbot"
+exit
+}
+
+#---------------------------------------------
+#                   is_file_installed
+#---------------------------------------------
+
+is_file_installed()
+{
+  local program=$1
+  
+  prognotfound=`$program -help | tail -1 | grep "not found" | wc -l`
+  if [ "$prognotfound" == "1" ] ; then
+    echo "***error: the program $program is not installed" 
+    return 1
+  fi
+  return 0
+}
+
+#---------------------------------------------
+#                   LIST_DESCENDANTS
+#---------------------------------------------
+
+LIST_DESCENDANTS ()
+{
+  local children=$(ps -o pid= --ppid "$1")
+
+  for pid in $children
+  do
+    LIST_DESCENDANTS "$pid"
+  done
+
+  echo "$children"
+}
+
+#VVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVVV
+#                             Primary script execution =
+#^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
 if [ ! -d ~/.cfastgit ] ; then
   mkdir ~/.cfastgit
 fi
@@ -23,48 +127,8 @@ else
   exit
 fi
 
-function usage {
-echo "Verification and validation testing script for cfast"
-echo ""
-echo "Options:"
-echo "-a - run automatically if cfast repo has changed"
-echo "-c - clean cfast and FDS-SMV repos"
-echo "-f - force cfastbot run"
-echo "-h - display this message"
-echo "-i - use installed smokeview and background (if using the 'none' queue)"
-echo "-I - compiler [ default: $compiler]"
-echo "-k - kill cfastbot"
-echo "-m email -  email_address "
-echo "-q queue_name - run cases using the queue queue_name"
-echo "     default: $QUEUE"
-echo "-s - skip matlab and guide generating stages"
-echo "-u - update cfast and FDS-SMV repos"
-echo "-U - upload guide (only by user: cfastbot)"
-echo "-v - show options used to run cfastbot"
-exit
-}
-
-is_file_installed()
-{
-  program=$1
-  prognotfound=`$program -help | tail -1 | grep "not found" | wc -l`
-  if [ "$prognotfound" == "1" ] ; then
-    echo "***error: the program $program is not installed" 
-    exit
-  fi
-}
-
-LIST_DESCENDANTS ()
-{
-  local children=$(ps -o pid= --ppid "$1")
-
-  for pid in $children
-  do
-    LIST_DESCENDANTS "$pid"
-  done
-
-  echo "$children"
-}
+botrepo=$reponame/bot
+botbranch=master
 
 RUNAUTO=
 UPDATEREPO=
@@ -81,6 +145,7 @@ havematlab=`which matlab 2> /dev/null | wc -l`
 UPLOAD=
 USEINSTALL=
 KILL_CFASTBOT=
+ECHO=
 
 while getopts 'acfhiI:km:q:suUv' OPTION
 do
@@ -126,6 +191,7 @@ case $OPTION  in
    ;;
   v)
    RUNCFASTBOT=0
+   ECHO=echo
    ;;
 esac
 done
@@ -138,15 +204,14 @@ if [ "$KILL_CFASTBOT" == "1" ]; then
     kill -9 $(LIST_DESCENDANTS $PID)
     echo "killing cfastbot (PID=$PID)"
     kill -9 $PID
-    JOBIDS=`qstat -a | grep CB_ | awk -v user="$USER" '{if($2==user){print $1}}'`
-    if [ "$JOBIDS" != "" ]; then
-      echo killing cfastbot jobs with Id:$JOBIDS
-      qdel $JOBIDS
+    if [ "$QUEUE" != "none"]; then
+      JOBIDS=`qstat -a | grep CB_ | awk -v user="$USER" '{if($2==user){print $1}}'`
+      if [ "$JOBIDS" != "" ]; then
+        echo killing cfastbot jobs with Id:$JOBIDS
+        qdel $JOBIDS
+      fi
     fi
     echo cfastbot process $PID killed
-    if [ -e $cfastbot_pid ]; then
-      rm $cfastbot_pid
-    fi
   else
     echo cfastbot is not running, cannot be killed.
   fi
@@ -156,8 +221,8 @@ fi
 if [ "$USEINSTALL" != "" ]; then
   echo
   echo looking for installed software
-  is_file_installed smokeview
-  is_file_installed background
+  is_file_installed smokeview || exit 1
+  is_file_installed background || exit 1
   echo "   found smokeview"
   echo "   found background"
 fi
@@ -176,10 +241,16 @@ fi
 if [[ "$UPDATEREPO" == "1" ]]; then
    UPDATEREPO=-u
    if [ "$RUNCFASTBOT" == "1" ]; then
-     cd $reponame/bot
-     git remote update
-     git checkout master
-     git merge origin/master
+     CD_REPO $botrepo $botbranch
+     git fetch origin
+     git merge origin/$botbranch
+
+     have_remote=`git remote -v | awk '{print $1}' | grep firemodels | wc  -l`
+     if [ "$have_remote" != "0" ]; then
+       git fetch firemodels
+       git merge firemodels/$botbranch
+     fi
+     
   fi
 fi
 if [[ "$CLEANREPO" == "1" ]]; then
@@ -198,9 +269,7 @@ QUEUE="-q $QUEUE"
 compiler="-I $compiler"
 PID="-p $cfastbot_pid"
 cd $CURDIR
-if [ "$RUNCFASTBOT" == "1" ] ; then
-  ./cfastbot.sh $PID $REPO $USEINSTALL $RUNAUTO $compiler $UPDATEREPO $CLEAN $QUEUE $SKIP $MATLABEXE $UPLOAD $EMAIL "$@"
-else
-  echo ./$botscript $PID $REPO $USEINSTALL $RUNAUTO $compiler $UPDATEREPO $CLEAN $QUEUE $SKIP $MATLABEXE $UPLOAD $EMAIL "$@"
-fi
-rm $cfastbot_pid
+$ECHO  ./cfastbot.sh $PID $REPO $USEINSTALL $RUNAUTO $compiler $UPDATEREPO $CLEAN $QUEUE $SKIP $MATLABEXE $UPLOAD $EMAIL "$@"
+if [ -e $cfastbot_pid]; then
+   rm $cfastbot_pid
+fi
